Create OneTimePadEncryption.py
This is a class file that can be import with from OneTimePadEncryption import *
The methods that you would call are the 
decrypt_file() --> which unzips the key.zip file that is generated with the encrypt_string_or_file() method
encrypt_string_or_file() --> which takes a string or a text file if string_file_mode is set to True and encrypts it with a randomly generated key, the key.dat file is actual zipped and protected with a password of your choice.  The key only works with the encrypted_message.txt that was generated at the same time.
decrypt_string_or_file() --> which can take the key was either a copy string, or read it directly from the key.dat file that was generated.  The encrypted string or the file the string is in.  You can choice whether you are using files by setting key_file_mode to True and or the encrypted_string_file_mode to True.
All other methods in this class are automatically called and therefore not called by the user.